pyproject.toml
[tool.poetry]
name = "django-assessment"
version = "0.1.0"
description = ""
authors = ["Deep Sarker <deepsarker01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
django-rest-framework = "^0.1.0"
mysqlclient = "^2.2.5"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
